--- chat-app/src/utils/database.js
@@ -1,15 +1,0 @@
-// src/utils/database.js
-const mongoose = require('mongoose');
-const connectDB = async () => {
-  try {
-    await mongoose.connect(process.env.MONGO_URI, {
-      useNewUrlParser: true,
-      useUnifiedTopology: true,
-    });
-    console.log('MongoDB connected');
-  } catch (err) {
-    console.error(err);
-    process.exit(1);
-  }
-};
-module.exports = connectDB;
